+++ apps/api/src/app/order/order.service.ts
@@ -1,16 +1,14 @@
 import { AccountService } from '@ghostfolio/api/app/account/account.service';
-import { CacheService } from '@ghostfolio/api/app/cache/cache.service';
 import { DataGatheringService } from '@ghostfolio/api/services/data-gathering.service';
 import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data.service';
 import { PrismaService } from '@ghostfolio/api/services/prisma.service';
 import { SymbolProfileService } from '@ghostfolio/api/services/symbol-profile.service';
 import {
-  DATA_GATHERING_QUEUE,
-  GATHER_ASSET_PROFILE_PROCESS
+  GATHER_ASSET_PROFILE_PROCESS,
+  GATHER_ASSET_PROFILE_PROCESS_OPTIONS
 } from '@ghostfolio/common/config';
 import { Filter } from '@ghostfolio/common/interfaces';
 import { OrderWithAccount } from '@ghostfolio/common/types';
-import { InjectQueue } from '@nestjs/bull';
 import { Injectable } from '@nestjs/common';
 import {
   AssetClass,
@@ -21,7 +19,6 @@ import {
   Type as TypeOfOrder
 } from '@prisma/client';
 import Big from 'big.js';
-import { Queue } from 'bull';
 import { endOfToday, isAfter } from 'date-fns';
 import { groupBy } from 'lodash';
 import { v4 as uuidv4 } from 'uuid';
@@ -32,11 +29,8 @@ import { Activity } from './interfaces/activities.interface';
 export class OrderService {
   public constructor(
     private readonly accountService: AccountService,
-    private readonly cacheService: CacheService,
-    @InjectQueue(DATA_GATHERING_QUEUE)
-    private readonly dataGatheringQueue: Queue,
-    private readonly exchangeRateDataService: ExchangeRateDataService,
     private readonly dataGatheringService: DataGatheringService,
+    private readonly exchangeRateDataService: ExchangeRateDataService,
     private readonly prismaService: PrismaService,
     private readonly symbolProfileService: SymbolProfileService
   ) {}
@@ -120,10 +114,14 @@ export class OrderService {
         data.SymbolProfile.connectOrCreate.create.symbol.toUpperCase();
     }
 
-    await this.dataGatheringQueue.add(GATHER_ASSET_PROFILE_PROCESS, {
-      dataSource: data.SymbolProfile.connectOrCreate.create.dataSource,
-      symbol: data.SymbolProfile.connectOrCreate.create.symbol
-    });
+    await this.dataGatheringService.addJobToQueue(
+      GATHER_ASSET_PROFILE_PROCESS,
+      {
+        dataSource: data.SymbolProfile.connectOrCreate.create.dataSource,
+        symbol: data.SymbolProfile.connectOrCreate.create.symbol
+      },
+      GATHER_ASSET_PROFILE_PROCESS_OPTIONS
+    );
 
     const isDraft = isAfter(data.date as Date, endOfToday());
 
@@ -137,8 +135,6 @@ export class OrderService {
         }
       ]);
     }
-
-    await this.cacheService.flush();
 
     delete data.accountId;
     delete data.assetClass;
@@ -329,8 +325,6 @@ export class OrderService {
         ]);
       }
     }
-
-    await this.cacheService.flush();
 
     delete data.assetClass;
     delete data.assetSubClass;