+++ apps/client/src/app/services/admin.service.ts
@@ -1,4 +1,4 @@
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { UpdateMarketDataDto } from '@ghostfolio/api/app/admin/update-market-data.dto';
 import { IDataProviderHistoricalResponse } from '@ghostfolio/api/services/interfaces/interfaces';
@@ -9,6 +9,7 @@ import {
   UniqueAsset
 } from '@ghostfolio/common/interfaces';
 import { DataSource, MarketData } from '@prisma/client';
+import { JobStatus } from 'bull';
 import { format, parseISO } from 'date-fns';
 import { Observable, map } from 'rxjs';
 
@@ -17,6 +18,22 @@ import { Observable, map } from 'rxjs';
 })
 export class AdminService {
   public constructor(private http: HttpClient) {}
+
+  public deleteJob(aId: string) {
+    return this.http.delete<void>(`/api/v1/admin/queue/job/${aId}`);
+  }
+
+  public deleteJobs({ status }: { status: JobStatus[] }) {
+    let params = new HttpParams();
+
+    if (status?.length > 0) {
+      params = params.append('status', status.join(','));
+    }
+
+    return this.http.delete<void>('/api/v1/admin/queue/job', {
+      params
+    });
+  }
 
   public deleteProfileData({ dataSource, symbol }: UniqueAsset) {
     return this.http.delete<void>(
@@ -43,16 +60,28 @@ export class AdminService {
       );
   }
 
-  public fetchJobs() {
-    return this.http.get<AdminJobs>(`/api/v1/admin/queue/jobs`);
+  public fetchJobs({ status }: { status?: JobStatus[] }) {
+    let params = new HttpParams();
+
+    if (status?.length > 0) {
+      params = params.append('status', status.join(','));
+    }
+
+    return this.http.get<AdminJobs>('/api/v1/admin/queue/job', {
+      params
+    });
+  }
+
+  public gather7Days() {
+    return this.http.post<void>('/api/v1/admin/gather', {});
   }
 
   public gatherMax() {
-    return this.http.post<void>(`/api/v1/admin/gather/max`, {});
+    return this.http.post<void>('/api/v1/admin/gather/max', {});
   }
 
   public gatherProfileData() {
-    return this.http.post<void>(`/api/v1/admin/gather/profile-data`, {});
+    return this.http.post<void>('/api/v1/admin/gather/profile-data', {});
   }
 
   public gatherProfileDataBySymbol({ dataSource, symbol }: UniqueAsset) {